Issue #3278469 by alex.ksis: Force adding #process and #pre_render callbacks for the social_tagging views filters.
The select2 plugin is not attached to the views filters provided by the social_tagging module if they are rendered using BEF.
If we just change the widget #type to select2, it won't work because these widgets will be rendered by the BEF plugin which adds the default #process and #pre_render callbacks. So in order to solve it, we need to force adding the correct callbacks which are provided by the select2 module.